integrated-video-analytics/requirements.txt
fastapi
httpx
uvicorn
python-multipart
opencv-python
torch
torchvision
numpy
pandas
requests
Pillow
deepface
tf-keras
easyocr
supervision
fpdf2
websockets
matplotlib>=3.3.0
pyyaml>=5.3.1
scipy>=1.4.1
tqdm>=4.64.0
psutil
seaborn>=0.11.0
ultralytics>=8.4.0
pytest>=8.0.0
